Add rubocop and fix cops

+++ lib/db/migrations/create_cards_colors_table.rb
@@ -1,17 +1,16 @@
+# frozen_string_literal: true
+
 class CreateCardsColorsTable < ActiveRecord::Migration[5.2]
-    def up
-      unless ActiveRecord::Base.connection.table_exists?(:cards_colors)
-        create_table :cards_colors, id: false do |table|
-          table.belongs_to :card
-          table.belongs_to :color
-        end
-      end
-    end
-  
-    def down
-      if ActiveRecord::Base.connection.table_exists?(:cards_colors)
-        drop_table :cards_colors
+  def up
+    unless ActiveRecord::Base.connection.table_exists?(:cards_colors)
+      create_table :cards_colors, id: false do |table|
+        table.belongs_to :card
+        table.belongs_to :color
       end
     end
   end
-  
+
+  def down
+    drop_table :cards_colors if ActiveRecord::Base.connection.table_exists?(:cards_colors)
+  end
+end
