+++ lib/db/migrations/create_colors_table.rb
@@ -1,17 +1,16 @@
+# frozen_string_literal: true
+
 class CreateColorsTable < ActiveRecord::Migration[5.2]
-    def up
-      unless ActiveRecord::Base.connection.table_exists?(:colors)
-        create_table :colors do |table|
-          table.string :name
-          table.timestamps
-        end
-      end
-    end
-  
-    def down
-      if ActiveRecord::Base.connection.table_exists?(:colors)
-        drop_table :colors
+  def up
+    unless ActiveRecord::Base.connection.table_exists?(:colors)
+      create_table :colors do |table|
+        table.string :name
+        table.timestamps
       end
     end
   end
-  
+
+  def down
+    drop_table :colors if ActiveRecord::Base.connection.table_exists?(:colors)
+  end
+end
